refactor: use handlers in place of multiple routers
